Fixes + Backup Organization
Fixed Automatic Update process by replacing unattended-upgrades with DietPi's script in /boot/dietpi.txt.

Additionally, I moved backups to ~/backups so the user running the script has access to files (not defaulting to 'dietpi') and they are out of the way.

Lastly, there is a fix in the intro prompt to state that this project uses README.md and not README.txt.